deployment/server/requirements.txt
# Web server
fastapi==0.95.0
uvicorn==0.21.1
gunicorn==20.1.0
python-multipart==0.0.6
pydantic==1.10.7

# Data processing
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
PyYAML==6.0

# Machine learning
tensorflow==2.12.0
torch==2.0.0
torchvision==0.15.1

# Storage & Monitoring
prometheus-client==0.16.0
python-dotenv==1.0.0
boto3==1.26.115
azure-storage-blob==12.16.0

# Testing
pytest==7.3.1
httpx==0.24.0